Add future physical-state prediction (state_future)
 Auxiliary Physical-State Predictor on System‑2 Intent Features

Default: This feature is off by default.
 1、Model Architecture
 - Input: mean‑pooled bridge tokens (from System‑2 intent features)
 - Processing: 2‑layer MLP
 - Output: predicts \( S_{t:t+H} \) (future states for H+1 frames)
- Loss: masked L2 loss computed in the normalized state space, kept “off the DiT action head” (pure bypass supervision)
 2、 Data Sampling
- Labels are sampled from state columns, reusing the original state normalization statistics.
 3、 CLI Switches & Trainer
- New CLI flags are added to enable the feature.
- Trainer sets `save_total_limit=0`, meaning **checkpoints are never auto‑deleted**.
4、 Fixes Included
- Removed stray `pdb.set_trace()` in evaluation wrappers.
- Fixed `grad_hook` dtype for bf16 precision.
- Added a fixer for the `episode‑701` dataset.
- Includes SF (state‑future) training scripts and `DIAL_state` documentation.
5、 Experiment Settings in This Repository
- Hardware: single RTX 4090 (48 GB)
- Batch size: `bs=4` with 16 accumulation steps (effective batch = 64)
- Precision & optimization: bf16 + gradient checkpointing
- Training pipeline:  
  （1）. 80k steps of “golden” pretraining (with `state_future` enabled)  
  （2）. 80k steps of end‑to‑end fine‑tuning
- Validation: evaluated on 24 RoboCasa GR1 tasks × 50 episodes each, compared against a 47.3% baseline.

+++ gr00t/model/backbone/vlm_backbone_dial.py
@@ -128,7 +128,7 @@ class VLMBackbone(nn.Module):
             def grad_hook(grad):
                 if self._embed_tokens_hook_mask.device != grad.device:
                     self._embed_tokens_hook_mask = self._embed_tokens_hook_mask.to(grad.device)
-                return grad * self._embed_tokens_hook_mask
+                return grad * self._embed_tokens_hook_mask.to(grad.dtype)
 
             self._embed_tokens_hook_handle = embed_tokens.weight.register_hook(grad_hook)
